app_proc/requirements.txt
aerospike
kafka-python